accidently implemented the feature in the wrong service, so moved it
