+++ src/main/java/com/group16b/ApiLayer/ProductionCompanyController.java
@@ -60,5 +60,11 @@ public class ProductionCompanyController extends BaseController {
     {
         return executeWithReturnData(() -> productionCompanyService.getCompanyAllEvents(companyId));
     }
+
+    @GetMapping("/{companyId}/permissions")
+    public ResponseEntity<?> getCompanyPerms(@PathVariable("companyId") int companyId)
+    {
+        return executeWithReturnData(() -> productionCompanyService.getCompanyAllEvents(companyId));
+    }
     
 }
